+++ src/main/java/chess/domain/piece/King.java
@@ -1,10 +1,8 @@
 package chess.domain.piece;
 
 import chess.domain.board.Board;
-import chess.domain.board.Cell;
 import chess.domain.board.Coordinate;
 import chess.domain.board.Direction;
-import chess.domain.player.TeamType;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,19 +18,14 @@ public class King extends Piece {
     public boolean isMovableTo(Board board, Coordinate currentCoordinate,
         Coordinate targetCoordinate) {
         Direction moveCommandDirection = currentCoordinate.calculateDirection(targetCoordinate);
-        List<Coordinate> possibleCoordinates = new ArrayList<>();
-        List<Direction> directions = getDirections();
-        if (!directions.contains(moveCommandDirection)) {
+        if (!isCorrectDirection(moveCommandDirection)) {
             return false;
         }
-
-        Coordinate movingCoordinate = currentCoordinate.move(moveCommandDirection);
-
-        Cell cell = board.find(movingCoordinate);
-        if (cell.isMovable(getTeamType())) {
-            possibleCoordinates.add(movingCoordinate);
+        Coordinate nextCoordinate = currentCoordinate.move(moveCommandDirection);
+        List<Coordinate> possibleCoordinates = new ArrayList<>();
+        if (board.isMovable(nextCoordinate, getTeamType())) {
+            possibleCoordinates.add(nextCoordinate);
         }
-
         return possibleCoordinates.contains(targetCoordinate);
     }
 }
