+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
 import chess.domain.board.Direction;
-import chess.domain.player.TeamType;
 
 public class Queen extends Piece {
     private static final String NAME = "Q";
