+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
 import chess.domain.board.Direction;
-import chess.domain.player.TeamType;
 
 public class Rook extends Piece {
     private static final String NAME = "R";
